Assignment 02
Q) Write a program in java that - 1) Prompts the user to enter their name, age, and whether they are a student or not. 2) It then reads the input using the Scanner class with appropriate data types (String, int, and boolean). 3) Based on the user's age, it determines if they are an adult or a minor, and based on their student status, it provides a relevant message. (eg: You are eligible to be a student etc). Note: Try to use your own method / functions to print messages.